Fix CI failures: docker build, codecov, e2e correctness
- Dockerfile: only enable the deadsnakes PPA for non-3.12 builds: both
  noble's own repo and deadsnakes provide python3.12 and mixing them
  breaks apt (all Python 3.12 lanes failed at image build)
- Drop codecov from CI: tokenless uploads are no longer accepted by
  the service (API lanes failed at the upload step)
- e2e executors: format command-line arguments at float64 precision;
  formatting float32 values at float32 precision produces a different
  double than the exact value, flipping tree-threshold comparisons
  (RandomForestRegressor failed on Java and JavaScript)
- RIPPER/IREP assembler: weight the per-rule constants by multiplying
  with the 0/1 match indicator instead of branching on it; strictly
  typed targets cannot use a numeric expression as an if-condition
  (wittgenstein models failed to compile on Java)

+++ tests/utils.py
@@ -398,7 +398,11 @@ def format_arg(value):
     if np.isnan(value):
         return "NaN"
 
-    return format_float(value)
+    # format at float64 precision: np.format_float_positional on a float32
+    # value produces its shortest float32 representation, which is a
+    # different double than the exact value and can flip tree-threshold
+    # comparisons in the generated code
+    return format_float(np.float64(value))
 
 
 def write_content_to_file(content, path):
